Fix compiler warnings

+++ src/ex/lf_buf_mngr.hpp
@@ -74,7 +74,7 @@ class LFBufMngr // : Buffer<T, Allocator>
      */
     template<typename _t = Buffer,
              std::enable_if_t<has_config<_t>::value, bool> = true>
-    LFBufMngr(int p_nbufs, int p_maxiters, typename _t::config_t config);
+    LFBufMngr(size_t p_nbufs, int p_maxiters, typename _t::config_t config);
     // ~LFBufMngr(){
     //   LOG(INFO)<<"D LFBuffer";
     // }
@@ -105,7 +105,7 @@ LFBufMngr<Buffer>::LFBufMngr(size_t p_nbufs, size_t p_buf_size, size_t p_max_tri
 template<class Buffer>
 template<typename _t,
          std::enable_if_t<has_config<_t>::value, bool>>
-LFBufMngr<Buffer>::LFBufMngr(int p_nbufs, int p_maxiters, typename _t::config_t config)
+LFBufMngr<Buffer>::LFBufMngr(size_t p_nbufs, int p_maxiters, typename _t::config_t config)
   : m_max_iters(p_maxiters)
   , m_nbufs(p_nbufs)
 {
